Add __construct__ function to the date dtype

+++ include/dynd/dtypes/strided_array_dtype.hpp
@@ -103,9 +103,9 @@ inline dtype make_strided_array_dtype(const dtype& element_dtype) {
     return dtype(new strided_array_dtype(element_dtype));
 }
 
-inline dtype make_strided_array_dtype(const dtype& uniform_dtype, int ndim) {
+inline dtype make_strided_array_dtype(const dtype& uniform_dtype, size_t ndim) {
     dtype result = uniform_dtype;
-    for (int i = 0; i < ndim; ++i) {
+    for (size_t i = 0; i < ndim; ++i) {
         result = make_strided_array_dtype(result);
     }
     return result;
